Refactor: timeStore -> pickDate
