+++ src/stores/timeStore.js
@@ -1,21 +1,9 @@
+import pickDate from '@/utils/pickDate'
 import { create } from 'zustand'
 
 const useTimeStore = create(set => ({
   dates: [],
-  selectDate: date =>
-    set(state => {
-      const dates = state.dates || [] // Ensure state.dates is an array
-      const dateString = date.toDateString()
-      if (dates.some(d => d.toDateString() === dateString)) {
-        return {
-          dates: dates.filter(d => d.toDateString() !== dateString),
-        }
-      } else {
-        return {
-          dates: [...dates, date],
-        }
-      }
-    }),
+  selectDate: date => set(state => pickDate(state, date)),
 }))
 
 export default useTimeStore
