+++ src/styles/timeSelect/TimeSelectPage.module.css
@@ -3,4 +3,7 @@
   height: 100%;
   margin: 0 auto;
   padding: 3.5% 0 1.5% 0;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
